mission_controller: add support for surfacing without waiting for ctrl-c/r
Add a new behavior (mode 6) to standard backse01.mi to surface with end_action
set to 2 (resume).

Add an argument to mission controller's surface behavior that switched between
the standard surface behavior and this new one.

Use x_surface_active and m_depth to track when we're on the surface due to a
surface instead of x_in_surface_dialog. I do not believe the surface dialog is
shown when end_action is 2 on surface behaviors.